+++ src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "../slices/auth";
-import messageReducer from "../slices/message";
+import authReducer from "./slices/auth";
+import messageReducer from "./slices/message";
 
 const reducer = {
   auth: authReducer,
